+++ web/src/state/app_action.rs
@@ -14,6 +14,6 @@ pub enum AppAction {
     SetAudioLoading(bool),
     SetSampleRate(u32),
     SetDensity(f32),
-    SetGrainLenMax(u32),
-    SetGrainLenMin(u32),
+    SetGrainLenMax(f32),
+    SetGrainLenMin(f32),
 }
